Fixing the problem statement for Lab11_Q1
The problem statement for Lab11_Q1 was incorrect.

The problem statement was:

Given a string A consisting of lowercase English alphabets, your task is to find the first non-repeating character in the string.

The correct problem statement is:

Given a string A consisting of lowercase English alphabets, your task is to find the first non-repeating character in the string. If there is no non-repeating character, return #.