Fixes #1
The problem was that the `Comparable` interface was not being implemented correctly.

The `Comparable` interface is a generic interface, and so it needs to be implemented with a type parameter.

The type parameter should be the type of the class implementing the interface.

The `Comparable` interface is also a functional interface, and so it can be implemented using a lambda expression.

The lambda expression should take an object of the type of the class implementing the interface, and return an integer.

The integer should be negative if the object is less than the object being compared to, zero if the objects are equal, and positive if the object is greater than the object being compared to.

The `getMax` function was also not implemented correctly.

The `Comparable` interface is a generic interface, and so it needs to be implemented with a type parameter.

The type parameter should be the type of the class implementing the interface.

The `Comparable` interface is also a functional interface, and so it can be implemented using a lambda expression.

The lambda expression should take an object of the type of the class implementing the interface, and return an integer.

The integer should be negative if the object is less than the object being compared to, zero if the objects are equal, and positive if the object is greater than the object being compared to.